% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies_fibos.R
\name{fibos_config}
\alias{fibos_config}
\title{Install the Python 'fibos' Module}
\usage{
fibos_config()
}
\description{
This function creates a Python virtual environment (on Windows) and installs the
Python module 'fibos' required for the full functionality of this package.
}
\note{
This function will install external software (a Python package) on your system.
}
