% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibos_execute.R
\name{osp_internal}
\alias{osp_internal}
\title{Calculate OSP Metric}
\usage{
osp_internal(file)
}
\arguments{
\item{file}{Path to the SRF file generated by the 'occluded_surface' function.}
}
\value{
A data frame containing the OSP metric data read from the generated PAK file.
}
\description{
Internal function to calculate the OSP metric from an SRF file.
This function calls the 'Python' implementation of the OSP calculation.
}
\author{
Carlos Henrique da Silveira (carlos.silveira@unifei.edu.br)

Herson Hebert Mendes Soares (hersonhebert@hotmail.com)

Joao Paulo Roquim Romanelli (joaoromanelli@unifei.edu.br)

Patrick Fleming (Pat.Fleming@jhu.edu)
}
\keyword{internal}
