% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respak.R
\name{read_osp}
\alias{read_osp}
\title{Read OSP Values}
\usage{
read_osp(prot_file)
}
\arguments{
\item{prot_file}{OSP File (.pak).}
}
\value{
A data frame containing the OSP data.
}
\description{
Internal function to read OSP values from a PAK file generated by the OSP calculation.
}
\keyword{internal}
