% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies_fibos.R
\name{fibos_config}
\alias{fibos_config}
\title{Install the 'Python' 'fibos' Module}
\usage{
fibos_config()
}
\value{
Invisible NULL. Called for its side effects.
}
\description{
This function creates a 'Python' virtual environment and installs the
'Python' module 'fibos' required for the full functionality of this package.
It handles different system configurations and ensures that the correct
compiler paths are set.
}
\note{
This function will install external software (a 'Python' package) on your system.
Administrator/sudo privileges might be required on some systems.
}
\examples{
\donttest{
# Set up the 'Python' environment and install the required module
fibos_config()
}
}
